fix some junit tests broken
Move the integration test for the draft restful endpoint to integration
package
